+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/BuildImageTask.java
@@ -91,7 +91,11 @@ public class BuildImageTask extends DefaultTask implements JibTask {
       AbsoluteUnixPath appRoot = TaskCommon.getAppRootChecked(jibExtension, getProject());
       GradleProjectProperties projectProperties =
           GradleProjectProperties.getForProject(
-              getProject(), getLogger(), jibExtension.getExtraDirectoryPath(), appRoot);
+              getProject(),
+              getLogger(),
+              jibExtension.getExtraDirectory().getPath(),
+              jibExtension.getExtraDirectory().getPermissions(),
+              appRoot);
       RawConfiguration rawConfiguration = new GradleRawConfiguration(jibExtension);
 
       if (Strings.isNullOrEmpty(jibExtension.getTo().getImage())) {