+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/BuildTarTask.java
@@ -84,7 +84,8 @@ public class BuildTarTask extends DefaultTask implements JibTask {
   @InputFiles
   public FileCollection getInputFiles() {
     return GradleProjectProperties.getInputFiles(
-        Preconditions.checkNotNull(jibExtension).getExtraDirectoryPath().toFile(), getProject());
+        Preconditions.checkNotNull(jibExtension).getExtraDirectory().getPath().toFile(),
+        getProject());
   }
 
   /**
@@ -119,7 +120,11 @@ public class BuildTarTask extends DefaultTask implements JibTask {
 
       GradleProjectProperties projectProperties =
           GradleProjectProperties.getForProject(
-              getProject(), getLogger(), jibExtension.getExtraDirectoryPath(), appRoot);
+              getProject(),
+              getLogger(),
+              jibExtension.getExtraDirectory().getPath(),
+              jibExtension.getExtraDirectory().getPermissions(),
+              appRoot);
       RawConfiguration rawConfiguration = new GradleRawConfiguration(jibExtension);
 
       GradleHelpfulSuggestionsBuilder gradleHelpfulSuggestionsBuilder =